fix story
