+++ libs/domain/content/link/link.model.ts
