+++ libs/domain/content/link/link.schema.ts
