--- libs/domain/content/link/src/index.ts
@@ -1,2 +1,0 @@
-export * from './link.component';
-export * from './link.model';
