+++ libs/domain/content/src/components.ts
@@ -1,3 +1,3 @@
 export * from '../banner/src/banner.def';
-export * from '../link/src/link.def';
+export * from '../link/link.def';
 export * from '../video/src/video.def';
